Make clear directory work with dir that doesn't exist

+++ src/main/java/io/domisum/lib/auxiliumlib/util/FileUtil.java
@@ -304,6 +304,8 @@ public final class FileUtil
 	{
 		try
 		{
+			if(!directory.exists())
+				return;
 			FileUtils.cleanDirectory(directory);
 		}
 		catch(IOException e)
